# RI_HFX| much better sparsity management in RHO flavor forces