This re-implementation of the forces in the RHO flavor of RI-HFX manages sparsity in a better way than before. In particular, the order of the many 3-center tensor contractions is changed such that densification happens at the last possible step, thus preserving sparsity for longer. This is partly made possible thanks to #1984.