@abussy 2022-03-16
From my own tests (on rather larger and sparser systems, without GPUs), I have observed a major performance benefit when it comes to the force evaluation. The overall time spent in the routine is at least halved. I didn't spent much time looking at more specific timings though. Having a closer look, I have for 64 water on 16 nodes (96 MPI ranks, 2 OMP threads per rank):

|Time (s)    | Old      | New    |
|--------------|-----------|----------|
| SCF         | 680      | 680     |
| Forces     | 575      | 141     |
| dbt_contract | 1041 |  656 |
| dbm_multiply | 772 | 461 |

So here I have a clear improvement, both int `dbt_contract` and `dbm_multiply`. I suspect that the 32 water system might be too small and not sparse enough for us to see the benefit, and that we have some overhead associated to `retain_sparsity`. I'll do an extra run with GPU to compare. I also agree that better understanding `dbt_contract` cannot hurt.