@oschuett 2022-03-16
Thanks for double-checking! Yes, these small single-node benchmarks always have to be taken with a grain of salt. Comparing the changes for the [MPI vs OpenMP](https://dashboard.cp2k.org/archive/perf-openmp/index.html) run is also interesting. I believe what we're seeing for `dbt_contract` is the lack of OpenMP parallelism combined with load-imbalance. That's good news because these problems should be fairly straightforward to fix.