@hfp 2022-03-16
> single-node benchmarks always have to be taken with a grain of salt.

Indeed, but still useful to set expectations for strong and weak scaling performance. Is there a chance to extend the miniapp to have (larger/different) inputs reproducing "real" performance like the task files for the grid backend?