@oschuett 2022-03-16
> Is there a chance to extend the miniapp to have (larger/different) inputs reproducing "real" performance like the task files for the grid backend?

Yes, this is definitely high up on my list.